move and rename file for new architecture
