+++ tests/Integration/inc/Engine/CDN/Subscriber/TestCase.php
@@ -1,18 +1,18 @@
 <?php
 
-namespace WP_Rocket\Tests\Integration\inc\classes\subscriber\CDN\CDNSubscriber;
+namespace WP_Rocket\Tests\Integration\inc\Engine\CDN\Subscriber;
 
 use WPMedia\PHPUnit\Integration\TestCase as BaseTestCase;
 use WP_Rocket\Admin\Options_Data;
 use WP_Rocket\Admin\Options;
-use WP_Rocket\CDN\CDN;
-use WP_Rocket\Subscriber\CDN\CDNSubscriber;
+use WP_Rocket\Engine\CDN\CDN;
+use WP_Rocket\Engine\CDN\Subscriber;
 
 abstract class TestCase extends BaseTestCase {
 
 	protected function getSubscriberInstance() {
 		$options = new Options_Data( ( new Options( 'wp_rocket_' ) )->get( 'settings' ) );
 
-		return new CDNSubscriber( $options, new CDN( $options ) );
+		return new Subscriber( $options, new CDN( $options ) );
 	}
 }
