+++ tests/Integration/inc/Engine/CDN/Subscriber/getCdnHosts.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace WP_Rocket\Tests\Integration\inc\classes\subscriber\CDN\CDNSubscriber;
+namespace WP_Rocket\Tests\Integration\inc\Engine\CDN\Subscriber;
 
 /**
- * @covers \WP_Rocket\Subscriber\CDN\CDNSubscriber::get_cdn_hosts
- * @uses   \WP_Rocket\CDN\CDN::get_cdn_urls
+ * @covers \WP_Rocket\Engine\CDN\Subscriber::get_cdn_hosts
+ * @uses   \WP_Rocket\Engine\CDN\CDN::get_cdn_urls
  * @uses   ::rocket_add_url_protocol
  * @group  Subscriber
  * @group  CDN
